// frontend/src/shared/api/generated/schemas/credentialPoolCredentialResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Bab API
 * OpenAPI spec version: 0.1.0
 */
import type { ProviderCredentialResponse } from './providerCredentialResponse';

export interface CredentialPoolCredentialResponse {
  id: string;
  org_id: string;
  pool_id: string;
  provider_credential_id: string;
  priority: number;
  weight: number;
  is_active: boolean;
  created_at: string;
  updated_at: string;
  credential: ProviderCredentialResponse;
}